% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adstock_w.R
\name{adstock_w}
\alias{adstock_w}
\title{Apply Adstock Weights}
\usage{
adstock_w(x, w, l)
}
\arguments{
\item{x}{vector}

\item{w}{weights}

\item{l}{maximum adstock length}
}
\value{
a numeric vector of length(x) with adstock transformed media impressions.
}
\description{
A function that will convolve over the one dimensional vector of raw impressions over time
and apply the adstock weights.
}
\examples{
data(mmm_imps)
x <- mmm_imps$mi_banners
w <- geometric_adstock_weights()
l <- length(w)
a <- adstock_w(x, w, l)

x[1:12]
a[1:12]

}
