% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_metrics.R
\name{channel_metrics}
\alias{channel_metrics}
\title{Channel Metrics}
\usage{
channel_metrics(model_decomp)
}
\arguments{
\item{model_decomp}{model decomposition table}
}
\description{
Channel Metrics
}
