% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve_adstock_weights.R
\name{convolve_adstock_weights}
\alias{convolve_adstock_weights}
\title{Convolve Adstock Weights}
\usage{
convolve_adstock_weights(x, w)
}
\arguments{
\item{x}{vector}

\item{w}{weights}
}
\value{
a numeric vector of length(x) with adstock transformed media impressions.
}
\description{
A function that will convolve over the one dimensional vector of raw impressions over time
and apply the adstock weights.
}
\examples{
data(mmm_imps)
x <- mmm_imps$mi_banners
w <- geometric_adstock_weights()
a <- convolve_adstock_weights(x, w)

x[1:12]
a[1:12]

}
