% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve_x.R
\name{convolve_x}
\alias{convolve_x}
\title{Convolve Adstock x}
\usage{
convolve_x(x, w, l)
}
\arguments{
\item{x}{vector}

\item{w}{weights}

\item{l}{len}
}
\value{
a numeric vector of length(x) with adstock transformed media impressions.
}
\description{
A function that will convolve over the one dimensional vector of raw impressions over time
and apply the adstock weights.
}
