% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Decompose Model Contributions}
\usage{
decompose(x, new_data, spend_data)
}
\arguments{
\item{x}{trained mmm model workflow}

\item{new_data}{model impression data}

\item{spend_data}{spend data}
}
\description{
Decompose Model Contributions
}
