% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed_adstock.R
\name{delayed_adstock}
\alias{delayed_adstock}
\title{Delayed Adstock Transformations}
\usage{
delayed_adstock(
  x,
  decay = 0.5,
  delay = 2,
  max_carryover = 12,
  normalize = TRUE
)
}
\arguments{
\item{decay}{decay rate}

\item{delay}{delay shift parameter}

\item{max_carryover}{Max Carryover}

\item{normalize}{should weights be normalized}
}
\description{
Delayed Adstock Transformations
}
