% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation_params.R
\name{extract_params}
\alias{extract_params}
\title{Extract step transformation parameters}
\usage{
extract_params(model_decomp, type = "step_hill_saturation")
}
\arguments{
\item{model_decomp}{model decomposition table}
}
\description{
Extract step transformation parameters
}
