% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_adstock.R
\name{geometric_adstock}
\alias{geometric_adstock}
\title{Geometric Adstock}
\usage{
geometric_adstock(x, decay = 0.5, max_carryover = 12, normalize = TRUE)
}
\arguments{
\item{x}{input vector}

\item{decay}{decay rate}

\item{max_carryover}{maximum carryover}

\item{normalize}{should the result be scaled?}
}
\description{
Geometric Adstock
}
