% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_adstock_weights.R
\name{geometric_adstock_weights}
\alias{geometric_adstock_weights}
\title{Geometric Adstock Weights}
\usage{
geometric_adstock_weights(decay = 0.5, max_carryover = 12, normalize = TRUE)
}
\arguments{
\item{decay}{decay rate}

\item{max_carryover}{maximum carryover}

\item{normalize}{should the result be scaled?}
}
\description{
Geometric Adstock Weights
}
