% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill_saturation.R
\name{hill_saturation}
\alias{hill_saturation}
\title{Hill Saturation}
\usage{
hill_saturation(
  x,
  half_saturation = NULL,
  shape = NULL,
  nu = 1,
  max_ref = FALSE
)
}
\arguments{
\item{x}{input vector}

\item{half_saturation}{half_saturation}

\item{shape}{shape}

\item{nu}{nu}

\item{max_ref}{should result be scale back to scale of x?}
}
\description{
Hill Saturation
}
