% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_metrics.R
\name{model_metrics}
\alias{model_metrics}
\title{Model Metrics}
\usage{
model_metrics(model_decomp, new_data)
}
\arguments{
\item{model_decomp}{model decomposition table}

\item{new_data}{model impressions data}
}
\description{
Model Metrics
}
