% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base_contribution.R
\name{plot_base_contribution}
\alias{plot_base_contribution}
\title{Plot Base Contribution}
\usage{
plot_base_contribution(model_decomp)
}
\arguments{
\item{model_decomp}{model decomposition table}
}
\description{
Plot Base Contribution
}
