% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_channel_contribution.R
\name{plot_channel_contribution}
\alias{plot_channel_contribution}
\title{Plot Channel Contribution}
\usage{
plot_channel_contribution(model_decomp)
}
\arguments{
\item{model_decomp}{model decomposition table}
}
\description{
Plot Channel Contribution
}
