% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_response_curves.R
\name{plot_response_curves}
\alias{plot_response_curves}
\title{user step function that tackes the recipe and adds to it}
\usage{
plot_response_curves(response_curves)
}
\description{
@param response_curves response curves data table
}
