% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curves.R
\name{response_curves}
\alias{response_curves}
\title{Generate Response Curves}
\usage{
response_curves(model_decomp)
}
\arguments{
\item{model_decomp}{model decomposition table/s}
}
\description{
Generate Response Curves
}
