% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{roi}
\alias{roi}
\title{ROI Calculation}
\usage{
roi(x, new_data, spend_data)
}
\arguments{
\item{x}{trained mmm model workflow}

\item{new_data}{impressions data}

\item{spend_data}{spend data}
}
\description{
ROI Calculation
}
