% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation_params.R
\name{saturation_params}
\alias{saturation_params}
\title{user step function that tackes the recipe and adds to it}
\usage{
saturation_params(model_decomp)
}
\arguments{
\item{model_decomp}{model decomposition table}
}
\description{
user step function that tackes the recipe and adds to it
}
