% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_hill_saturation.R
\name{step_hill_saturation}
\alias{step_hill_saturation}
\title{user step function that tackes the recipe and adds to it}
\usage{
step_hill_saturation(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  half_saturation = NULL,
  shape = 1,
  nu = 1,
  max_ref = FALSE,
  columns = NULL,
  skip = FALSE,
  id = recipes::rand_id("hill_saturation")
)
}
\arguments{
\item{recipe}{recipe}

\item{role}{role}

\item{trained}{trained or not}

\item{half_saturation}{half_saturation}

\item{shape}{shape}

\item{nu}{nu}

\item{max_ref}{max_ref}

\item{columns}{columns}

\item{skip}{FALSE}

\item{id}{id}
}
\description{
user step function that tackes the recipe and adds to it
}
