% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymmm-package.R
\docType{package}
\name{tidymmm-package}
\alias{tidymmm}
\alias{tidymmm-package}
\title{tidymmm: Tools For Tidy Marketing Mix Modeling}
\description{
Package of MMM tools that work with the tidymodels framework and post modeling tools for model evaluation and optimization.
}
\keyword{internal}
