% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mmm_dataset.R
\name{validate_mmm_dataset}
\alias{validate_mmm_dataset}
\title{Validate media investment and spend datasets}
\usage{
validate_mmm_dataset(imps_df, spend_df)
}
\arguments{
\item{imps_df}{media investment data}

\item{spend_df}{media spend data}
}
\description{
Validate media investment and spend datasets
}
