% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_adstock.R
\name{weibull_adstock}
\alias{weibull_adstock}
\title{Weibull Adstock Transformations}
\usage{
weibull_adstock(x, shape = 1, scale = 1, max_carryover = 12, normalize = TRUE)
}
\arguments{
\item{shape}{decay rate}

\item{scale}{delay shift parameter}

\item{max_carryover}{Max Carryover}

\item{normalize}{should weights be normalized}
}
\description{
Weibull Adstock Transformations
}
