% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_adstock_weights.R
\name{weibull_adstock_weights}
\alias{weibull_adstock_weights}
\title{Weibull Adstock Weights}
\usage{
weibull_adstock_weights(
  shape = 1,
  scale = 1,
  max_carryover = 12,
  normalize = TRUE
)
}
\arguments{
\item{shape}{weibull shape parameter}

\item{scale}{weibull scale parameter}

\item{max_carryover}{max_carryover parameter}

\item{normalize}{should the result be scaled?}
}
\value{
a numeric vector of weights of length (max_carryover)
}
\description{
weibull_adstock_weights is a function that is called to generate a vector of weights
that define how media impressions will be decayed in present and carried over into the future.
}
\examples{
weibull_adstock_weights()
weibull_adstock_weights(shape = 0.5, scale = 10)
weibull_adstock_weights(shape = 1.5, scale = 20)

}
