% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adstock_w.R
\name{adstock_w}
\alias{adstock_w}
\title{Apply Adstock Weights}
\usage{
adstock_w(x, w, l)
}
\arguments{
\item{x}{vector}

\item{w}{weights}

\item{l}{maximum adstock length}
}
\description{
Apply Adstock Weights
}
