% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_adstock.R
\name{geometric_adstock}
\alias{geometric_adstock}
\title{Geometric Adstock}
\usage{
geometric_adstock(x, decay = 0.5, max_carryover = 12, normalize = TRUE)
}
\arguments{
\item{x}{input vector}

\item{decay}{decay rate}

\item{max_carryover}{maximum carryover}

\item{normalize}{should the result be scaled?}
}
\value{
a numeric vector of length(x) with adstock transformed media impressions.
}
\description{
A function to apply the geometric adstock transfromation to a vector of media
impressions
}
\examples{
data(mmm_imps)
x <- mmm_imps$mi_banners
a <- geometric_adstock(x)
b <- geometric_adstock(x, decay = 0.7, max_carryover = 2)

x[1:12]
a[1:12]
b[1:12]

plot(x[1:20], type = "l", main = "Geometric Adstock Transformation of x")
lines(a[1:20], col = 2)
lines(b[1:20], col = 3)


}
