% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_adstock_weights.R
\name{geometric_adstock_weights}
\alias{geometric_adstock_weights}
\title{Geometric Adstock Weights}
\usage{
geometric_adstock_weights(decay = 0.5, max_carryover = 12, normalize = TRUE)
}
\arguments{
\item{decay}{decay rate}

\item{max_carryover}{maximum carryover}

\item{normalize}{should the result be scaled?}
}
\value{
a numeric vector of weights of length (max_carryover)
}
\description{
geometric_adstock_weights is a function that is called to generate a vector of weights
that define how media impressions will be decayed in present and carried over into the future.
}
\examples{
geometric_adstock_weights()
geometric_adstock_weights(decay = 0.1)
geometric_adstock_weights(max_carryover = 3)
geometric_adstock_weights(decay = 0.4, max_carryover = 4, normalize = FALSE)

}
