% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill_saturation.R
\name{hill_saturation}
\alias{hill_saturation}
\title{Hill Saturation}
\usage{
hill_saturation(
  x,
  half_saturation = NULL,
  shape = NULL,
  nu = 1,
  max_ref = FALSE
)
}
\arguments{
\item{x}{input vector}

\item{half_saturation}{half_saturation}

\item{shape}{shape}

\item{nu}{nu}

\item{max_ref}{should result be scale back to scale of x?

#'@examples
data(mmm_imps)
x <- mmm_imps$mi_banners
y <- hill_saturation(x, shape = 1)
y_scaled <- hill_saturation(x, shape = 1, max_ref = TRUE)

x[1:12]
y[1:12]
y_scaled[1:12]

plot(sort(x), sort(y), type = "p", main = 'Hill Saturated x (scaled to [0-1])')
plot(sort(x), sort(y_scaled), type="p", main = "Scaled to Back to scale of x")
points(sort(x), sort(x), col = 2)}
}
\value{
a numeric vector of length(x) with saturated media impressions.
}
\description{
A function to apply the Hill Saturation to a vector of media impressions
}
