% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curves.R
\name{response_curves}
\alias{response_curves}
\title{user step function that tackes the recipe and adds to it}
\usage{
response_curves(model_decomp)
}
\description{
@param model_decomp model decomposition table
}
