% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_geometric_adstock.R
\name{step_geometric_adstock}
\alias{step_geometric_adstock}
\title{Step Geometric Adstock}
\usage{
step_geometric_adstock(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  decay = 0.5,
  max_carryover = 12,
  columns = NULL,
  skip = FALSE,
  id = recipes::rand_id("geometric_adstock")
)
}
\arguments{
\item{recipe}{recipe}

\item{role}{role}

\item{trained}{trained or not}

\item{decay}{decay rate}

\item{max_carryover}{max_carryover}

\item{columns}{columns}

\item{skip}{FALSE}

\item{id}{id}
}
\description{
user step function that takes the recipe and adds to it
}
